+++ tests/HslTest.php
@@ -11,10 +11,10 @@ class HslTest extends TestCase
 {
     public function test__construct()
     {
-        $obj = new Hsl(25, 25, 25);
-        $this->assertEquals(0.0980392157, round($obj->getHue(), 10));
-        $this->assertEquals(0.0980392157, round($obj->getSaturation(), 10));
-        $this->assertEquals(0.0980392157, round($obj->getLightness(), 10));
+        $obj = new Hsl(90, 0.25, 0.25);
+        $this->assertEquals(0.25, round($obj->getHue(), 10));
+        $this->assertEquals(0.25, round($obj->getSaturation(), 10));
+        $this->assertEquals(0.25, round($obj->getLightness(), 10));
         $this->assertEquals(1.0, $obj->getAlpha());
     }
 
@@ -26,8 +26,8 @@ class HslTest extends TestCase
         $this->assertEquals('hsl(180, 10%, 25%)', $color->toCss(true));
         $this->assertEquals('#394646', $color->toCssHex());
         $color->setAlpha(0.2);
-        $this->assertEquals('hsla(180 10% 23.75% / 0.2)', $color->toCss());
-        $this->assertEquals('hsla(180, 10%, 23.75%, 0.2)', $color->toCss(true));
+        $this->assertEquals('hsla(180 10% 25% / 0.2)', $color->toCss());
+        $this->assertEquals('hsla(180, 10%, 25%, 0.2)', $color->toCss(true));
     }
 
     public function testFromRgb()